document_service/requirements.txt
asgiref==3.9.1
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.1
lxml==6.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyPDF2==3.0.1
pytesseract==0.3.13
python-decouple==3.8
python-docx==1.2.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2